# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Fersaku API
  version: 0.0.1
  description: "Fersaku hosted storefront and QRIS payment gateway API. Transport contract (BE-110): success envelope, problem\
    \ errors, cursor meta. Identity/session lifecycle (BE-120): cookie sessions, Argon2id, CSRF, MFA. Account profile/security/preferences\
    \ (BE-125): profile versioning, password change, dual-confirm email change, MFA disable, notification preferences. RBAC/tenant\
    \ authorization (BE-130): permission registry, system roles, merchant membership, RequirePermission middleware; sample\
    \ admin/seller/buyer authz routes. Roles/assignments/invitations (BE-135): custom role CRUD, anti-escalation, staff/merchant\
    \ invitations with §6.5 POST body token accept (GET cannot consume). Notification inbox/dispatch foundation (BE-140):\
    \ recipient-scoped inbox list/read/read-all/badge, closed event registry aligned with BE-125 mandatory/optional prefs,\
    \ outbox email/dispatch jobs, CTA sanitization (relative internal only). Buyer/admin/seller notification paths are thin\
    \ aliases. Merchant/store onboarding (BE-200): mandatory canonical store, slug normalize/availability, idempotent POST\
    \ /v1/onboarding/store, progress patch, complete without product (ADR-0007). Fee policy/value objects (BE-300): pure int64\
    \ whole-IDR calculators (3% half-up + Rp700; withdrawal 3% + provider; min Rp50k), checksum-verified LAUNCH_FEE_POLICY_V1\
...
    \ FE client ops.\n"
servers:
...
paths:
...
  /v1/seller/me/merchant:
    get:
      operationId: sellerMeMerchant
      summary: Seller bootstrap — merchant, memberships, stores, current store (INT-150)
      description: |
        Returns primary merchant, all active memberships with store-scoped capabilities,
        allowed stores, canonicalStoreId, and server-selected currentStoreId.
        Selection: valid preferred store → canonical → first stable store.
        Cross-tenant store IDs are never returned. Membership required (403 if none).
      tags:
      - seller
      security:
      - sessionCookie: []
      responses:
        '200':
          description: Seller bootstrap snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '403':
          description: No active merchant membership
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemEnvelope'
  /v1/seller/me/current-store:
    put:
      operationId: sellerSetCurrentStore
      summary: Persist preferred current store (must be membership-owned)
      tags:
      - seller
      security:
      - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - storeId
              properties:
                storeId:
                  type: string
      responses:
        '200':
          description: Updated current store selection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '404':
          description: Unknown or cross-tenant store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemEnvelope'
...
  /v1/buyer/resources/{ownerUserId}:
    get:
      operationId: buyerResourceProbe
      summary: Buyer ownership probe sample (mismatch → RESOURCE_NOT_FOUND)
      tags:
      - buyer
      security:
      - sessionCookie: []
      parameters:
      - name: ownerUserId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Caller owns resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '404':
          description: Not owner (cross-tenant policy)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemEnvelope'
...
  /v1/checkout/stock-reservations:
    post:
      operationId: reserveStock
      summary: Atomic stock unit reservation (checkout foundation; FOR UPDATE SKIP LOCKED)
      tags:
      - inventory
      - checkout
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - storeId
              - productId
              properties:
                storeId:
                  type: string
                productId:
                  type: string
                orderId:
                  type: string
                checkoutId:
                  type: string
      responses:
        '201':
          description: Reserved (masked item only)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryReserveEnvelope'
        '200':
          description: Idempotent replay
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryReserveEnvelope'
        '409':
          description: INVENTORY_OUT_OF_STOCK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemEnvelope'
      security: []
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: fersaku_session
      description: HttpOnly opaque browser session cookie. Required for authenticated browser/SSR surfaces. Pair unsafe methods
        with X-CSRF-Token when CSRF is enforced.
    merchantApiKey:
      type: http
      scheme: bearer
      bearerFormat: fsk_
      description: 'Merchant gateway API key (Authorization: Bearer fsk_test_* sandbox / fsk_live_* live). Payment-only scope
        under /v1/gateway and deprecated /v1/qris. Never cookie, query, or public FE env. LIVE create requires ACTIVE QRIS_API
        capability (KYC).'
    xenditCallbackToken:
      type: apiKey
      in: header
      name: X-Callback-Token
      description: Inbound Xendit provider callback token. Constant-time verified before any business mutation. Not a session
        cookie and not a merchant API key.
    capabilityToken:
      type: apiKey
      in: header
      name: X-Capability-Token
      description: 'Optional capability/scoped token for non-session public flows (checkout owner proof, delivery access).
        Distinct from sessionCookie and merchantApiKey. When absent, operation may still allow public access via empty security:
        [] where declared.'
...
  schemas:
...
    Meta:
      type: object
      required:
      - requestId
      - timestamp
      description: Base success meta. Pagination fields live on CursorListMeta or NumberedPageListMeta only.
      properties:
        requestId:
          type: string
        timestamp:
          type: string
          format: date-time
          description: RFC3339 UTC timestamp
          example: '2026-07-17T10:00:00Z'
    SuccessEnvelope:
      type: object
      required:
      - data
      - meta
      description: Canonical success wire envelope. data is operation-specific; meta is Meta or a list profile.
      properties:
        data: {}
        meta:
          $ref: '#/components/schemas/Meta'
    Problem:
      type: object
      required:
      - code
      - message
      - requestId
      properties:
        code:
          type: string
          description: Stable machine code (§6.3)
          example: VALIDATION_FAILED
        message:
          type: string
          description: Safe client message (no secrets/PII/stack)
        details:
          type: object
          description: Optional structured details; field errors under details.fields for VALIDATION_FAILED
          properties:
            fields:
              type: array
              items:
                $ref: '#/components/schemas/FieldViolation'
          additionalProperties: true
        requestId:
          type: string
    ProblemEnvelope:
      type: object
      required:
      - problem
      description: Canonical problem wire envelope.
      properties:
        problem:
          $ref: '#/components/schemas/Problem'
...
    InventoryStockItemMasked:
      type: object
      required:
      - id
      - productId
      - storeId
      - schemaVersion
      - status
      - masked
      description: List/export DTO — secrets always masked; never encrypted_payload
      properties:
        id:
          type: string
        productId:
          type: string
        storeId:
          type: string
        schemaVersion:
          type: integer
        status:
          type: string
          enum:
          - AVAILABLE
          - RESERVED
          - DELIVERED
          - REVOKED
        masked:
          type: object
          additionalProperties:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    InventoryReserveEnvelope:
      type: object
      required:
      - data
      - meta
      properties:
        data:
          type: object
          required:
          - reservation
          - item
          - replayed
          properties:
            reservation:
              type: object
              properties:
                id:
                  type: string
                stockItemId:
                  type: string
                productId:
                  type: string
                status:
                  type: string
                expiresAt:
                  type: string
                  format: date-time
            item:
              $ref: '#/components/schemas/InventoryStockItemMasked'
            replayed:
              type: boolean
        meta:
          $ref: '#/components/schemas/Meta'
...
    FieldViolation:
      type: object
      required:
      - field
      - code
      properties:
        field:
          type: string
        code:
          type: string
          example: INVALID
        message:
          type: string
...
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Client-stable UUID/key for safe mutation replay within TTL.
      schema:
        type: string
        maxLength: 128
...
security: []